enhance: device pages for user
