+++ ui/user/src/lib/services/user/types.ts
@@ -229,6 +229,102 @@ export interface BootstrapStatus {
 	enabled: boolean;
 }
 
+// Devices
+
+// Device scans — payload shape matches apiclient/types/devicescan.go.
+export interface DeviceScan {
+	id: number;
+	receivedAt: string;
+	submittedBy?: string;
+	scannerVersion: string;
+	scannedAt: string;
+	deviceID: string;
+	hostname: string;
+	os: string;
+	arch: string;
+	username?: string;
+	files: DeviceScanFile[];
+	mcpServers: DeviceScanMCPServer[];
+	skills: DeviceScanSkill[];
+	plugins: DeviceScanPlugin[];
+	clients: DeviceScanClient[];
+}
+export interface DeviceScanClient {
+	name: string;
+	version?: string;
+	binaryPath?: string;
+	installPath?: string;
+	configPath?: string;
+	hasMCPServers: boolean;
+	hasSkills: boolean;
+	hasPlugins: boolean;
+}
+export interface DeviceScanFile {
+	path: string;
+	sizeBytes: number;
+	oversized: boolean;
+	content?: string;
+}
+export interface DeviceScanList {
+	items: DeviceScan[] | null;
+}
+export type DeviceScanListFilters = {
+	limit?: number;
+	offset?: number;
+	submittedBy?: string[];
+	deviceId?: string[];
+	groupByDevice?: boolean;
+};
+export interface DeviceScanMCPServer {
+	id: number;
+	client: string;
+	projectPath?: string;
+	file?: string;
+	configHash?: string;
+	envKeys: string[];
+	headerKeys: string[];
+	name: string;
+	transport: string;
+	command?: string;
+	args?: string[];
+	url?: string;
+}
+export interface DeviceScanPlugin {
+	id: number;
+	client: string;
+	projectPath?: string;
+	configPath?: string;
+	name: string;
+	pluginType: string;
+	version?: string;
+	description?: string;
+	author?: string;
+	marketplace?: string;
+	files: string[];
+	enabled: boolean;
+	hasMCPServers: boolean;
+	hasSkills: boolean;
+	hasRules: boolean;
+	hasCommands: boolean;
+	hasHooks: boolean;
+}
+export interface DeviceScanResponse extends DeviceScanList {
+	total: number;
+	limit: number;
+	offset: number;
+}
+export interface DeviceScanSkill {
+	id: number;
+	client: string;
+	projectPath?: string;
+	file?: string;
+	name: string;
+	description?: string;
+	files: string[];
+	hasScripts: boolean;
+	gitRemoteURL?: string;
+}
+
 // Files
 
 export interface Files {
