+++ ui/user/src/routes/devices/[device_id]/+page.ts
@@ -1,17 +1,16 @@
 import { handleRouteError } from '$lib/errors';
-import { AdminService } from '$lib/services';
-import type { DeviceScanResponse } from '$lib/services/admin/types';
-import { profile } from '$lib/stores';
+import { UserService, type DeviceScanResponse } from '$lib/services';
 import type { PageLoad } from './$types';
 
 const PAGE_SIZE = 50;
 
-export const load: PageLoad = async ({ params, fetch }) => {
+export const load: PageLoad = async ({ params, fetch, parent }) => {
+	const { profile } = await parent();
 	const { device_id } = params;
 
 	let scans: DeviceScanResponse = { items: [], total: 0, limit: PAGE_SIZE, offset: 0 };
 	try {
-		scans = await AdminService.listDeviceScans(
+		scans = await UserService.listDeviceScans(
 			{
 				limit: PAGE_SIZE,
 				deviceId: [device_id],
@@ -21,6 +20,7 @@ export const load: PageLoad = async ({ params, fetch }) => {
 		);
 		return { scans, deviceId: device_id, pageSize: PAGE_SIZE };
 	} catch (err) {
-		handleRouteError(err, `/admin/devices/${device_id}`, profile.current);
+		const prefix = profile.hasAdminAccess?.() ? '/admin' : '';
+		handleRouteError(err, `${prefix}/devices/${device_id}`, profile);
 	}
 };
