+++ ui/user/src/routes/devices/[device_id]/scans/[scan_id]/mcp/[id]/+page.svelte
@@ -6,7 +6,8 @@
 	import Layout from '$lib/components/Layout.svelte';
 	import { PAGE_TRANSITION_DURATION } from '$lib/constants';
 	import { deriveDeviceScope, formatDeviceClient } from '$lib/format.js';
-	import type { DeviceScanMCPServer } from '$lib/services/admin/types';
+	import type { DeviceScanMCPServer } from '$lib/services/user/types';
+	import { profile } from '$lib/stores';
 	import { goto } from '$lib/url';
 	import { findParentPlugin, shortHash } from '../../_shared/files';
 	import { fly } from 'svelte/transition';
@@ -17,7 +18,11 @@
 	let server = $derived<DeviceScanMCPServer | undefined>(
 		scan?.mcpServers?.find((m) => m.id === id)
 	);
-	let backHref = $derived(`/admin/devices/${page.params.device_id}/scans/${page.params.scan_id}`);
+	let hasAdminAccess = $derived(profile.current.hasAdminAccess?.());
+	let urlPrefix = $derived((hasAdminAccess ? '/admin' : '') as `/${string}`);
+	let backHref = $derived(
+		`${urlPrefix}/devices/${page.params.device_id}/scans/${page.params.scan_id}`
+	);
 
 	let endpoint = $derived(
 		server
@@ -143,7 +148,7 @@
 							<a
 								class="text-sm text-link"
 								href={resolve(
-									`/admin/devices/${page.params.device_id}/scans/${page.params.scan_id}/plugins/${parentPlugin.id}`
+									`${urlPrefix}/devices/${page.params.device_id}/scans/${page.params.scan_id}/plugins/${parentPlugin.id}`
 								)}
 							>
 								{parentPlugin.name}
